ernie-allsenses/requirements.txt
# ERNIE AI Developer Challenge - AllSensesAI Guardian
# Python dependencies for demo application

# Core web framework
Flask==3.0.0
Flask-CORS==4.0.0

# ERNIE API client (OpenAI-compatible)
openai>=1.40.0

# Utilities
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-flask==1.3.0
requests==2.31.0

# Production server (optional)
gunicorn==21.2.0